refactor: RubikCubeGroupの生成メソッドをstatic メソッドに移動
