+++ workspaces/cube-clarity-core/lib/three/animator/3d/rotatedCubeState.ts
@@ -16,7 +16,7 @@ import {
   type RubikCubePieceState,
 } from "../RubikCubeAnimator.ts";
 import { enumerate, flatMap, map } from "@core/iterutil/pipe";
-import { RUBIK_CUBE_PIECES } from "../../RubikCubeModel.ts";
+import { RUBIK_CUBE_PIECES } from "../../RubikCubeGroup.ts";
 import { rotateByWorldAxis } from "./helper.ts";
 
 export const ROTATION_INFORMATIONS = {
