+++ workspaces/cube-clarity-core/lib/three/mod.ts
@@ -1,3 +1,3 @@
-export * from "./RubikCubeModel.ts";
+export * from "./RubikCubeGroup.ts";
 export * from "./animator/RubikCubeAnimator.ts";
 export * from "./animator/MoveNotationTween.ts";
